+++ packages/toolboxes/utxo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swapkit/toolbox-utxo
 
+## 1.0.0-rc.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@1.0.0-rc.48
+
 ## 1.0.0-rc.50
 
 ### Patch Changes