+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swapkit/wallet-evm-extensions
 
+## 1.0.0-rc.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/toolbox-evm@1.0.0-rc.51
+
 ## 1.0.0-rc.50
 
 ### Patch Changes